--- sneakers/lib/services/repository/icons_services.dart
@@ -1,1 +1,0 @@
-
