+++ sneakers/lib/view/pages/Home/home_page.dart
@@ -5,12 +5,14 @@ import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:google_fonts/google_fonts.dart';
 import 'package:nepseapp/model/item.dart';
 import 'package:nepseapp/repository/icons_services.dart';
+import 'package:nepseapp/repository/itemservices.dart';
 import 'package:nepseapp/view/widgets/UI/homepage_ui/home_page_ui.dart';
 import 'package:nepseapp/view_model/home_Item/Brands/bloc/home_bloc.dart';
 
 class HomePage extends StatelessWidget {
   final IconsServices iconsServices;
-  const HomePage({Key? key, required this.iconsServices}) : super(key: key);
+  final ItemServices itemServices;
+  const HomePage({Key? key, required this.iconsServices,required this.itemServices}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
@@ -19,98 +21,11 @@ class HomePage extends StatelessWidget {
       padding: EdgeInsets.only(top: max_size.height * 0.01),
       child: BlocProvider(
         create: (context) =>
-            HomeBloc(iconsServices: iconsServices)..add(ItemLoading()),
+            HomeBloc(iconsServices: iconsServices, itemServices:itemServices )..add(ItemLoading()),
         child: const HomePageUI(),
       ),
     );
   }
 }
 
-//This UI woulde be used for the bottom of the page
 
-class CatlogItems extends StatelessWidget {
-  final Item item;
-  const CatlogItems({Key? key, required this.item}) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return Padding(
-      padding: const EdgeInsets.only(left: 12.0, right: 12.0, top: 10.0),
-      child: Container(
-        decoration: BoxDecoration(
-            color: Theme.of(context).cardColor,
-            borderRadius: BorderRadius.circular(16.0)),
-        child: Column(
-          children: [
-            Container(
-              decoration: const BoxDecoration(
-                  borderRadius: BorderRadius.only(
-                      topLeft: Radius.circular(16.0),
-                      topRight: Radius.circular(16.0))),
-              height: MediaQuery.of(context).size.height * 0.2,
-              width: MediaQuery.of(context).size.width * 0.5,
-              child: Image.asset(
-                item.imageUrl,
-                fit: BoxFit.cover,
-              ),
-            ),
-            Text(
-              item.name,
-              style: TextStyle(
-                  color: Theme.of(context).canvasColor,
-                  fontWeight: FontWeight.bold),
-            ),
-
-            //! Heart and the Buy button Widget container
-            const FooterContainer()
-          ],
-        ),
-      ),
-    );
-  }
-}
-
-class FooterContainer extends StatelessWidget {
-  const FooterContainer({Key? key}) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return Row(
-      children: [
-        IconButton(
-            onPressed: () {},
-            icon: const Icon(
-              CupertinoIcons.heart,
-              color: Colors.red,
-            )),
-        Padding(
-          padding: const EdgeInsets.only(left: 39.0),
-          child: Material(
-            borderRadius: BorderRadius.circular(16.0),
-            child: InkWell(
-              borderRadius: BorderRadius.circular(16.0),
-              onTap: () {},
-              child: Ink(
-                decoration: BoxDecoration(
-                  borderRadius: BorderRadius.circular(16.0),
-                  color: Colors.red,
-                ),
-                child: Center(
-                  child: Text(
-                    "Buy",
-                    style: TextStyle(
-                        color: Colors.white,
-                        fontWeight: FontWeight.bold,
-                        fontFamily: GoogleFonts.saira().fontFamily),
-                  ),
-                ),
-                height: MediaQuery.of(context).size.height * 0.04,
-                width: MediaQuery.of(context).size.width * 0.18,
-              ),
-            ),
-          ),
-        )
-      ],
-    );
-  }
-}
