+++ sneakers/lib/view/widgets/UI/homepage_ui/home_page_ui.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:nepseapp/view/widgets/UI/homepage_ui/flash_sale.dart';
+import 'package:nepseapp/view/widgets/UI/homepage_ui/recent_view.dart';
 import 'package:nepseapp/view/widgets/UI/homepage_ui/search_bar.dart';
 
 import 'package:nepseapp/view/widgets/UI/homepage_ui/brands_Icons.dart';
@@ -12,16 +13,19 @@ class HomePageUI extends StatelessWidget {
   Widget build(BuildContext context) {
     var size = MediaQuery.of(context).size;
     return Material(
-      child: SizedBox(
-        height: size.height,
-        width: size.width,
-        child: Column(
-          children: const [
-            SearchBarUI(),
-            OfferCard(),
-            BrandsIconCard(),
-            FlashSale(),
-          ],
+      child: SingleChildScrollView(
+        child: SizedBox(
+          height: size.height,
+          width: size.width,
+          child: Column(
+            children: const [
+              SearchBarUI(),
+              OfferCard(),
+              BrandsIconCard(),
+              FlashSale(),
+              RecentlyViewed()
+            ],
+          ),
         ),
       ),
     );
